+++ spec_rails/dummy_app/spec/integration_spec.rb
@@ -2,12 +2,12 @@
 
 require "spec_helper"
 
-RSpec.describe "Axn::Openapi inside a Rails application" do
+RSpec.describe "Axn::OpenAPI inside a Rails application" do
   it "boots the dummy Rails app" do
     expect(Rails.application).to be_a(Rails::Application)
   end
 
   it "loads the gem" do
-    expect(defined?(Axn::Openapi)).to eq("constant")
+    expect(defined?(Axn::OpenAPI)).to eq("constant")
   end
 end